requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
